Sharkcano, the shark and the volcano.

This is just a free image off the internetz. The image was read in and all non-essential items dropped. The dimensions in raster::raster terms is stored in attr(sharkcano, "rasterdim").

Format

A data frame with 117843 rows and 2 variables:

cell_

integer, cell index

byte

integer, byte value of shark image pixels

These are cell values on a grid that is 648x958.

References

This is the small version from here, see script in data-raw/sharkcano.r http://www.freestockphotos.biz/stockphoto/16214 Thanks to @jennybc for pointers on finding free stuff: https://github.com/jennybc/free-photos

Examples

library(raster)
rd <- attr(sharkcano, "rasterdim")
rastershark <- raster(matrix(NA_integer_, rd[1], rd[2]))
rastershark[sharkcano$cell_] <- sharkcano$byte  ## byte, heh
## I present to you, Sharkcano!  (Just wait for the 3D version, Quadshark).
plot(rastercano)

#> Error in plot.new(): figure margins too large

contour(rastershark, add = TRUE, labels = FALSE)

#> Error in contour.default(x = xFromCol(x, 1:ncol(x)), y = yFromRow(x, nrow(x):1),     z = t((getValues(x, format = "matrix"))[nrow(x):1, ]), ...): plot.new has not been called yet

plot(rastershark, col = "black")

#> Error in plot.xy(xy, type, ...): invalid graphics state

## another way
plot(rastercano)

#> Error in plot.xy(xy, type, ...): invalid graphics state

points(xyFromCell(rastershark, sharkcano$cell_), pch = ".")

#> Error in plot.xy(xy.coords(x, y), type = type, ...): invalid graphics state