Tabular tools for raster

Extract and index with raster tidy tools for raster.

Details

Tabularaster includes these main functions.

as_tibble	convert raster data to data frame form, with control over output and form of dimension/coordinate columns

cellnumbers	extract a data frame of query identifiers and cell,pixel index numbers

decimate	fast and loose resizing of a raster to coarser resolution

index_extent	build an extent in row column form, as opposed to coordinate value form