Extract cell numbers from a Raster object.

Provide the 'cellnumbers' capability of raster::extract and friends directly, returning a data frame of query-object identifiers 'object_' and the cell number.

cellnumbers(x, query, ...)

Arguments

x
Raster object

query
Spatial object or matrix of coordinates

...
arguments passed on to raster::cellFromPolygon for polygon input

Value

tbl_df data frame

Details

Raster data is inherently 2-dimensional, with a time or 'level' dimension treated like a layers of these 2D forms. The 'raster' package cell number is counted from 1 at the top-left, across the rows and down. This corresponds the the standard "raster graphics" convention used by 'GDAL' and the 'sp' package, and many other implementations. Note that this is different to the convention used by the graphics::image function.

Currently this function only operats as if the input is a single layer objects, it's not clear if adding an extra level of grouping for layers would be sensible.

The dots argument can be used to set weights=TRUE for the polygon case, this is otherwise ignored.

Examples

library(raster)
r <- raster(volcano) %>% aggregate(fact = 4)
cellnumbers(r, rasterToContour(r, level = 120))

#> Warning: projections not the same 
#>     x: NA
#> query: NA

#> # A tibble: 64 x 2
#>    object_ cell_
#>      <int> <dbl>
#>  1       1     8
#>  2       1     9
#>  3       1    10
#>  4       1    11
#>  5       1    21
#>  6       1    22
#>  7       1    23
#>  8       1    24
#>  9       1    27
#> 10       1    28
#> # ... with 54 more rows

#library(dplyr)

#cr <- cut(r,  pretty(values(r)))
#p <- raster::rasterToPolygons(cr, dissolve = TRUE)
#p <- spex::qm_rasterToPolygons_sp(cr)
#tt <- cellnumbers(cr, p)
#library(dplyr)
#tt %>% mutate(v = extract(r, cell_)) %>% 
#group_by(object_) %>% 
#summarize(mean(v)) 
#head(pretty(values(r)), -1)