Extract cell numbers from a Raster object.

Currently only for single layer objects.

cellnumbers(x, query, ...)

Arguments

x
Raster object

query
Spatial object or matrix of coordinates

...
arguments passed on to raster::cellFromPolygon for polygon input

Value

tbl_df data frame

Details

The dots argument can be used to set weights=TRUE for the polygon case, this is otherwise ignored.

Examples

library(raster)
r <- raster(volcano) %>% aggregate(fact = 4)

#> Error in raster(volcano) %>% aggregate(fact = 4): could not find function "%>%"

cr <- cut(r,  pretty(values(r)))

#> Error in cut(r, pretty(values(r))): object 'r' not found

p <- raster::rasterToPolygons(cr, dissolve = TRUE)

#> Error in nlayers(x): object 'cr' not found

tt <- cellnumbers(cr, p)

#> Error in inherits(query, "sf"): object 'p' not found

library(dplyr)

#> Error in value[[3L]](cond): Package ‘dplyr’ version 0.7.0 cannot be unloaded:
#>  Error in unloadNamespace(package) : namespace ‘dplyr’ is imported by ‘tabularaster’, ‘quadmesh’, ‘spbabel’ so cannot be unloaded

tt %>% mutate(v = extract(r, cell_)) %>%
group_by(object_) %>%
summarize(mean(v))

#> Error in tt %>% mutate(v = extract(r, cell_)) %>% group_by(object_) %>%     summarize(mean(v)): could not find function "%>%"

head(pretty(values(r)), -1)

#> Error in values(r): object 'r' not found